requirements.txt
annotated-types==0.6.0
anyio==4.4.0
appdirs==1.4.4
bcrypt==4.0.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
data==0.4
fastapi==0.110.0
Flask==3.0.2
frozendict==2.4.4
h11==0.14.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.2
MarkupSafe==2.1.5
matplotlib==3.8.3
multitasking==0.0.11
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pandas-datareader==0.10.0
passlib==1.7.4
peewee==3.17.6
plotly==5.22.0
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.9.0
pymongo==4.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.36.3
wheel
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0ll
yfinance==0.2.38